Initial huge OOP refactoring effort
