+++ AttackGlobals.py
@@ -1,0 +1,1 @@
+ATK_TGT_UNKNOWN, ATK_TGT_SINGLE, ATK_TGT_MULTI = (0, 1, 2)  # previously 1,2,3
